Solve Day 3 #1
