+++ 2022/day2/src/strategy_guide.rs
@@ -5,61 +5,51 @@ use crate::moves::Move;
 
 #[derive(Debug)]
 pub struct StrategyGuide {
-    moves: Vec<(Move, Move)>
+    moves: Vec<(Move, Move)>,
 }
 
 impl StrategyGuide {
     pub fn new(file_contents: &String) -> StrategyGuide {
-
         let mut moves: Vec<(Move, Move)> = Vec::new();
 
         for line in file_contents.lines() {
             moves.push(Self::get_moves_from_guide_line(line))
         }
 
-        StrategyGuide {
-            moves
-        }
-
+        StrategyGuide { moves }
     }
 
     pub fn correct_new(file_contents: &String) -> StrategyGuide {
-
         let mut moves: Vec<(Move, Move)> = Vec::new();
 
         for line in file_contents.lines() {
             moves.push(Self::get_correct_moves_from_guide_line(line))
         }
 
-        StrategyGuide {
-            moves
-        }
-
+        StrategyGuide { moves }
     }
 
     pub fn get_total_score(&self) -> u64 {
-        
         let mut total_score = 0;
 
         for moves in self.moves.iter() {
             match moves {
-
                 // Draws
-                (Move::Rock, Move::Rock) => total_score += 3 + 1,  // Draw + Rock
+                (Move::Rock, Move::Rock) => total_score += 3 + 1, // Draw + Rock
                 (Move::Paper, Move::Paper) => total_score += 3 + 2, // Draw + Paper
                 (Move::Scissors, Move::Scissors) => total_score += 3 + 3, // Draw + Scissors
-                
+
                 // Wins
-                (Move::Scissors, Move::Rock) => total_score += 6 + 1,  // Win + Rock
-                (Move::Rock, Move::Paper) => total_score += 6 + 2,  // Win + Paper
+                (Move::Scissors, Move::Rock) => total_score += 6 + 1, // Win + Rock
+                (Move::Rock, Move::Paper) => total_score += 6 + 2,    // Win + Paper
                 (Move::Paper, Move::Scissors) => total_score += 6 + 3, // Win + Scissors
 
                 // Losses
-                (Move::Paper, Move::Rock) => total_score += 0 + 1,  // Loss + Rock
-                (Move::Scissors, Move::Paper) => total_score += 0 + 2,  // Loss + Paper
+                (Move::Paper, Move::Rock) => total_score += 0 + 1, // Loss + Rock
+                (Move::Scissors, Move::Paper) => total_score += 0 + 2, // Loss + Paper
                 (Move::Rock, Move::Scissors) => total_score += 0 + 3, // Loss + Scissors
 
-                _ => {},
+                _ => {}
             }
         }
 
@@ -70,17 +60,16 @@ impl StrategyGuide {
     // so use this helper function instead for whole line
 
     fn get_moves_from_guide_line(guide_line: &str) -> (Move, Move) {
-        
         let mut first_move = Move::Invalid;
         let mut second_move = Move::Invalid;
 
         let mut chars = guide_line.chars();
-        
+
         match chars.next() {
             Some('A') => first_move = Move::Rock,
             Some('B') => first_move = Move::Paper,
             Some('C') => first_move = Move::Scissors,
-            _   => {},
+            _ => {}
         }
 
         _ = chars.next();
@@ -89,25 +78,23 @@ impl StrategyGuide {
             Some('X') => second_move = Move::Rock,
             Some('Y') => second_move = Move::Paper,
             Some('Z') => second_move = Move::Scissors,
-            _   => {},
+            _ => {}
         }
 
         (first_move, second_move)
     }
 
-
     fn get_correct_moves_from_guide_line(guide_line: &str) -> (Move, Move) {
-        
         let mut first_move = Move::Invalid;
         let mut second_move = Move::Invalid;
 
         let mut chars = guide_line.chars();
-        
+
         match chars.next() {
             Some('A') => first_move = Move::Rock,
             Some('B') => first_move = Move::Paper,
             Some('C') => first_move = Move::Scissors,
-            _   => {},
+            _ => {}
         }
 
         _ = chars.next();
@@ -121,7 +108,7 @@ impl StrategyGuide {
                     Move::Scissors => second_move = Move::Paper,
                     _ => {}
                 }
-            },
+            }
             Some('Y') => {
                 // Draw
                 match first_move {
@@ -130,7 +117,7 @@ impl StrategyGuide {
                     Move::Scissors => second_move = Move::Scissors,
                     _ => {}
                 }
-            },
+            }
             Some('Z') => {
                 // Win
                 match first_move {
@@ -139,11 +126,10 @@ impl StrategyGuide {
                     Move::Scissors => second_move = Move::Rock,
                     _ => {}
                 }
-            },
-            _   => {},
+            }
+            _ => {}
         }
 
         (first_move, second_move)
     }
-
 } // impl StrategyGuide
