+++ 2022/y2022/src/main.rs
@@ -26,5 +26,4 @@ fn main() {
         "Day 3, #1: {}",
         day3::total_priorities_of_common_types(&String::from("resources/input/day03-1.txt"))
     );
-
 }
